agents-mcp/pyproject.toml
[project]
name = "agents-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "openai-agents>=0.0.8",
    "pydotenv>=0.0.7",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "sendgrid>=6.11.0",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.34.0",
]
